Split useColumns
